+++ R/darkgrey.R
@@ -1,5 +1,0 @@
-#  from usethis https://github.com/r-lib/usethis/
-
-darkgrey <- function(x) {
-  x <- crayon::make_style("darkgrey")(x)
-}