+++ R/utils.R
@@ -1,3 +1,8 @@
+#  from usethis https://github.com/r-lib/usethis/
+darkgrey <- function(x) {
+  x <- crayon::make_style("darkgrey")(x)
+}
+
 check_file_exist <- function(file){
   res <- TRUE
   if (file.exists(file)){