+++ inst/shinyexample/R/onload.R
@@ -1,8 +1,0 @@
-#' Adds the content of www to www/ from this package
-#'
-#' @importFrom shiny addResourcePath
-#'
-#' @noRd
-.onLoad <- function(...) {
-  shiny::addResourcePath('www', system.file('app/www', package = 'shinyexample'))
-}