+++ inst/shinyexample/dev/02_dev.R
@@ -6,8 +6,9 @@
 # 2. All along your project
 
 ## 2.1 Add modules
-
-golem::add_module( name = "my_other_module") # Name of the module 
+## 
+golem::add_module(name = "my_first_module") # Name of the module
+golem::add_module( name = "my_other_module") # Name of the module
 
 ## 2.2 Add dependencies
 
@@ -16,6 +17,16 @@ usethis::use_package("pkg") # To call each time you need a new package
 ## 2.3 Add tests
 
 usethis::use_test("app")
+
+## 2.4 Add a browser button
+
+golem::add_browser_button()
+
+## 2.5 Add external files
+
+golem::add_js_file("script")
+golem::add_js_handler("script")
+golem::add_css_file("custom")
 
 # 3. Documentation
 
@@ -27,3 +38,7 @@ devtools::build_vignettes()
 usethis::use_travis()
 usethis::use_appveyor()
 usethis::use_coverage()
+
+# You're now set! 
+# go to dev/03_deploy.R
+rstudioapi::navigateToFile("dev/03_deploy.R")